% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_fv.R
\name{annuity_fv}
\alias{annuity_fv}
\title{Future Value (FV) of an Annuity Function}
\usage{
annuity_fv(rate, nper, pmt)
}
\arguments{
\item{rate}{The discount rate.}

\item{nper}{The number of periods.}

\item{pmt}{The payment amount per period.}
}
\value{
The Future Value.
}
\description{
Calculates the Future Value (FV) of an annuity.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example 1: Future Value of an annuity with a 5\% discount rate over 10 periods and $100 payment per period
annuity_fv(0.05, 10, 100)

# Example 2: Future Value of an annuity with a 3\% discount rate over 15 periods and $200 payment per period
}
