% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_pv.R
\name{annuity_pv}
\alias{annuity_pv}
\title{Present Value (PV) of an Annuity Function}
\usage{
annuity_pv(rate, nper, pmt)
}
\arguments{
\item{rate}{The discount rate.}

\item{nper}{The number of periods.}

\item{pmt}{The payment amount per period.}
}
\value{
The Present Value.
}
\description{
Calculates the Present Value (PV) of an annuity.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example 1: Present Value of an annuity with a 5\% discount rate over 10 periods and $100 payment per period

# Example 2: Present Value of an annuity with a 3\% discount rate over 15 periods and $200 payment per period
}
