% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdy.R
\name{bdy}
\alias{bdy}
\title{Computing Bank Discount Yield (BDY)}
\usage{
bdy(d, f, t)
}
\arguments{
\item{d}{The dollar discount, which is equal to the difference between the face value of the bill and the purchase price.}

\item{f}{The face value (par value) of the bill.}

\item{t}{Number of days remaining until maturity.}
}
\value{
The bank discount yield as a decimal.
}
\description{
Computes the bank discount yield (BDY) for a T-bill.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
