% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdy2mmy.R
\name{bdy2mmy}
\alias{bdy2mmy}
\title{Computing Money Market Yield (MMY)}
\usage{
bdy2mmy(bdy, t)
}
\arguments{
\item{bdy}{The bank discount yield.}

\item{t}{Number of days remaining until maturity.}
}
\value{
The money market yield as a decimal.
}
\description{
Computes the money market yield (MMY) for a T-bill.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
