% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_price.R
\name{bond_price}
\alias{bond_price}
\title{Bond Price Function}
\usage{
bond_price(face_value, coupon_rate, market_rate, nper)
}
\arguments{
\item{face_value}{The face value (par value) of the bond.}

\item{coupon_rate}{The annual coupon rate of the bond.}

\item{market_rate}{The current market interest rate.}

\item{nper}{The number of periods (years) until bond maturity.}
}
\value{
The price of the bond.
}
\description{
Calculates the price of a bond given its face value, coupon rate, market rate, and number of periods.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
}
