% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_yield.R
\name{bond_yield}
\alias{bond_yield}
\title{Bond Yield Function}
\usage{
bond_yield(face_value, price, coupon_rate, nper)
}
\arguments{
\item{face_value}{The face value (par value) of the bond.}

\item{price}{The current market price of the bond.}

\item{coupon_rate}{The annual coupon rate of the bond.}

\item{nper}{The number of periods (years) until bond maturity.}
}
\value{
The bond yield.
}
\description{
Calculates the yield (rate of return) of a bond given its price, face value, coupon rate, and number of periods.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
