% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_yield_plotly.R
\name{bond_yield_plotly}
\alias{bond_yield_plotly}
\title{Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica}
\usage{
bond_yield_plotly(face_value, price, coupon_rate, nper)
}
\arguments{
\item{face_value}{The face value (par value) of the bond.}

\item{price}{The current market price of the bond.}

\item{coupon_rate}{The annual coupon rate of the bond.}

\item{nper}{The number of periods (years) until bond maturity.}
}
\value{
A plotly object representing the bond yield function.
}
\description{
Bond Yield Function Plot using Plotly
}
\details{
This function plots the bond yield function given the bond parameters.
}
\examples{
# Example usage:
}
