% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_irr.R
\name{calculate_irr}
\alias{calculate_irr}
\title{Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica}
\usage{
calculate_irr(initial_investment, cash_flows)
}
\arguments{
\item{initial_investment}{A numeric value representing the initial investment.}

\item{cash_flows}{A numeric vector of cash flows.}
}
\value{
A numeric value representing the internal rate of return.
}
\description{
Calculate Internal Rate of Return (IRR)
}
\details{
This function calculates the internal rate of return of a series of cash flows given an initial investment.
}
