% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mirr.R
\name{calculate_mirr}
\alias{calculate_mirr}
\title{#' Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica}
\usage{
calculate_mirr(initial_investment, cash_flows, finance_rate, reinvestment_rate)
}
\arguments{
\item{initial_investment}{A numeric value representing the initial investment.}

\item{cash_flows}{A numeric vector of cash flows.}

\item{finance_rate}{A numeric value representing the finance rate.}

\item{reinvestment_rate}{A numeric value representing the reinvestment rate.}
}
\value{
A numeric value representing the modified internal rate of return.
}
\description{
Calculate Modified Internal Rate of Return (MIRR)
}
\details{
This function calculates the modified internal rate of return (MIRR) of a series of cash flows given an initial investment, a finance rate, and a reinvestment rate.
}
\examples{
initial_investment <- -1000
cash_flows <- c(100, 200, 300, 400, 500)
finance_rate <- 0.05
reinvestment_rate <- 0.08
calculate_mirr(initial_investment, cash_flows, finance_rate, reinvestment_rate)
}
