% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_npv.R
\name{calculate_npv}
\alias{calculate_npv}
\title{Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica}
\usage{
calculate_npv(initial_investment, cash_flows, rate)
}
\arguments{
\item{initial_investment}{A numeric value representing the initial investment.}

\item{cash_flows}{A numeric vector of cash flows.}

\item{rate}{A numeric value representing the discount rate.}
}
\value{
A numeric value representing the net present value of the cash flows.
}
\description{
Calculate Net Present Value (NPV)
}
\details{
This function calculates the net present value of a series of cash flows given an initial investment and a discount rate.
}
\examples{
initial_investment <- -1000
cash_flows <- c(100, 200, 300)
rate <- 0.05
calculate_npv(initial_investment, cash_flows, rate)
}
