% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candlestickChart.R
\name{candlestickChart}
\alias{candlestickChart}
\title{Candlestick Chart}
\usage{
candlestickChart(ohlc, start = NULL, end = NULL, main = "", ...)
}
\arguments{
\item{ohlc}{Data frame with columns Date, Open, High, Low, Close.}

\item{start}{Start date to plot, if not specified, all dates in ohlc will be included.}

\item{end}{End date to plot.}

\item{main}{An overall title for the plot.}

\item{...}{Additional arguments to be passed to ggplot.}
}
\value{
A ggplot object showing the candlestick chart.
}
\description{
Shows prices for each period as a continuous line. The box is clear if the closing price is higher than the opening price, or filled red if the closing is lower than the opening price.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
candlestickChart(data, main = "Candlestick Chart Example")
}
