% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash.ratio.R
\name{cash.ratio}
\alias{cash.ratio}
\title{Cash Ratio}
\usage{
cash.ratio(cash, ms, cl)
}
\arguments{
\item{cash}{Cash.}

\item{ms}{Marketable securities.}

\item{cl}{Current liabilities.}
}
\value{
The cash ratio as a decimal.

The cash ratio as a decimal.
}
\description{
Computes the cash ratio – Liquidity ratios measure the firm’s ability to satisfy its short-term obligations as they come due.

Computes the cash ratio – Liquidity ratios measure the firm’s ability to satisfy its short-term obligations as they come due.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica

# Example usage:
cash.ratio(cash = 3000, ms = 2000, cl = 2000)
}
