% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash_flow_diagram.R
\name{cash_flow_diagram_ggplot2}
\alias{cash_flow_diagram_ggplot2}
\title{Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica}
\usage{
cash_flow_diagram_ggplot2(cash_flows, periods)
}
\arguments{
\item{cash_flows}{A numeric vector of cash flows.}

\item{periods}{A numeric vector of periods corresponding to the cash flows.}
}
\value{
A ggplot object representing the cash flow diagram.
}
\description{
Plot Cash Flow Diagram using ggplot2
}
\details{
This function plots an interactive cash flow diagram given a series of cash flows and periods.
}
