% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient.variation.R
\name{coefficient.variation}
\alias{coefficient.variation}
\title{Computing Coefficient of Variation}
\usage{
coefficient.variation(sd, avg)
}
\arguments{
\item{sd}{Standard deviation.}

\item{avg}{Average value.}
}
\value{
The coefficient of variation as a decimal.
}
\description{
Computes the coefficient of variation.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
