% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.ratio.R
\name{current.ratio}
\alias{current.ratio}
\title{Current Ratio}
\usage{
current.ratio(ca, cl)
}
\arguments{
\item{ca}{Current assets.}

\item{cl}{Current liabilities.}
}
\value{
The current ratio as a decimal.
}
\description{
Computes the current ratio – Liquidity ratios measure the firm’s ability to satisfy its short-term obligations as they come due.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
current.ratio(ca = 8000, cl = 2000)
}
