% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddb.R
\name{ddb}
\alias{ddb}
\title{Depreciation Expense Recognition – Double-Declining Balance}
\usage{
ddb(cost, rv, t)
}
\arguments{
\item{cost}{Cost of long-lived assets.}

\item{rv}{Residual value of the long-lived assets at the end of its useful life. DDB does not explicitly use the asset’s residual value in the calculations, but depreciation ends once the estimated residual value has been reached. If the asset is expected to have no residual value, the DB method will never fully depreciate it, so the DB method is typically changed to straight-line at some point in the asset’s life.}

\item{t}{Length of the useful life.}
}
\value{
A data frame showing the depreciation expense for each year.
}
\description{
Computes depreciation expense recognition using the double-declining balance (DDB) method, which applies two times the straight-line rate to the declining balance.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
