% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddm}
\alias{ddm}
\title{Dividend Discount Model (DDM) Function}
\usage{
ddm(dividends, discount_rate)
}
\arguments{
\item{dividends}{A numeric vector of expected future dividends.}

\item{discount_rate}{The discount rate (required rate of return).}
}
\value{
The estimated fair value of the stock.
}
\description{
Calculates the present value of future dividends using the Dividend Discount Model.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
#'
}
\examples{
# Example usage:
dividends <- c(2.5, 2.7, 3.0, 3.2, 3.5)
discount_rate <- 0.1
ddm(dividends, discount_rate)
}
