% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt.ratio.R
\name{debt.ratio}
\alias{debt.ratio}
\title{Debt Ratio}
\usage{
debt.ratio(td, ta)
}
\arguments{
\item{td}{Total debt.}

\item{ta}{Total assets.}
}
\value{
The debt ratio as a decimal.
}
\description{
Computes the debt ratio – Solvency ratios measure the firm’s ability to satisfy its long-term obligations.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
