% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diluted.EPS.R
\name{diluted.EPS}
\alias{diluted.EPS}
\title{Diluted Earnings Per Share}
\usage{
diluted.EPS(ni, pd, cpd = 0, cdi = 0, tax = 0, w, cps = 0, cds = 0, iss = 0)
}
\arguments{
\item{ni}{Net income.}

\item{pd}{Preferred dividends.}

\item{cpd}{Dividends on convertible preferred stock.}

\item{cdi}{Interest on convertible debt.}

\item{tax}{Tax rate.}

\item{w}{Weighted average number of common shares outstanding.}

\item{cps}{Shares from conversion of convertible preferred stock.}

\item{cds}{Shares from conversion of convertible debt.}

\item{iss}{Shares issuable from stock options.}
}
\value{
Diluted earnings per share.
}
\description{
Computes diluted earnings per share.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
diluted.EPS(ni = 115600, pd = 10000, cdi = 42000, tax = 0.4, w = 200000, cds = 60000)
diluted.EPS(ni = 115600, pd = 10000, cpd = 10000, w = 200000, cps = 40000)
diluted.EPS(ni = 115600, pd = 10000, w = 200000, iss = 2500)
}
