% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.rate.R
\name{discount.rate}
\alias{discount.rate}
\title{Discount Rate Calculation}
\usage{
discount.rate(n, pv, fv, pmt, type = 0)
}
\arguments{
\item{n}{Number of periods.}

\item{pv}{Present value.}

\item{fv}{Future value.}

\item{pmt}{Payment per period.}

\item{type}{0 for payments at the end of each period; 1 for payments at the beginning of each period.}
}
\value{
Rate of return for each period.
}
\description{
Computes the rate of return for each period based on given parameters.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
\seealso{
fv.simple fv.annuity fv pv pmt n.period
}
