% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear.continuous.R
\name{ear.continuous}
\alias{ear.continuous}
\title{Continuous Effective Annual Rate (EAR)}
\usage{
ear.continuous(r)
}
\arguments{
\item{r}{Stated annual rate.}
}
\value{
Effective annual rate with continuous compounding.
}
\description{
Converts a stated annual rate to the effective annual rate with continuous compounding.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
ear.continuous(r = 0.1)
}
