% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear.R
\name{ear}
\alias{ear}
\title{Effective Annual Rate (EAR)}
\usage{
ear(r, m)
}
\arguments{
\item{r}{Stated annual rate.}

\item{m}{Number of compounding periods per year.}
}
\value{
Effective annual rate.
}
\description{
Converts a stated annual rate to the effective annual rate.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
ear(r = 0.12, m = 12)
}
