% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear2bey.R
\name{ear2bey}
\alias{ear2bey}
\title{Bond-Equivalent Yield (BEY)}
\usage{
ear2bey(ear)
}
\arguments{
\item{ear}{Effective annual rate (EAR).}
}
\value{
Bond-equivalent yield (BEY).
}
\description{
Converts the effective annual rate (EAR) to the bond-equivalent yield (BEY), which is 2 times the semiannual discount rate.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
