% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ear2hpr.R
\name{ear2hpr}
\alias{ear2hpr}
\title{Computing HPR, the holding period return}
\usage{
ear2hpr(ear, t)
}
\arguments{
\item{ear}{Effective annual rate.}

\item{t}{Number of days remaining until maturity.}
}
\value{
Holding period return (HPR).
}
\description{
Computes the holding period return (HPR) based on the effective annual rate (EAR) and the number of days remaining until maturity.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
\seealso{
hpr2ear ear hpr
}
