% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIR.R
\name{EIR}
\alias{EIR}
\title{Equivalent/proportional Interest Rates}
\usage{
EIR(r, n = 1, p = 12, type = c("e", "p"))
}
\arguments{
\item{r}{Interest rate to be applied n times per year (r is annual rate!).}

\item{n}{Times that the interest rate r is compounded per year.}

\item{p}{Times that the equivalent rate is compounded per year.}

\item{type}{Equivalent interest rates ('e', default) or proportional interest rates ('p').}
}
\value{
Equivalent or proportional interest rate.
}
\description{
An interest rate to be applied n times per year can be converted to an equivalent rate to be applied p times per year.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
EIR(r=0.05, n=1, p=12)
EIR(r=0.05, n=2, p=12)
EIR(r=0.05, n=4, p=12)
EIR(r=0.05, n=12, p=1)
EIR(r=0.05, n=1, p=4)

}
\seealso{
ear
}
