% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPS.R
\name{EPS}
\alias{EPS}
\title{Basic Earnings Per Share}
\usage{
EPS(ni, pd, w)
}
\arguments{
\item{ni}{Net income.}

\item{pd}{Preferred dividends.}

\item{w}{Weighted average number of common shares outstanding.}
}
\value{
Basic earnings per share (EPS).
}
\description{
Computes the basic earnings per share (EPS) based on net income, preferred dividends, and the weighted average number of common shares outstanding.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
\seealso{
diluted.EPS
}
