% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial.leverage.R
\name{financial.leverage}
\alias{financial.leverage}
\title{Financial Leverage}
\usage{
financial.leverage(te, ta)
}
\arguments{
\item{te}{Total equity.}

\item{ta}{Total assets.}
}
\value{
Financial leverage ratio.
}
\description{
Computes the financial leverage, which is a solvency ratio that measures the firm's ability to satisfy its long-term obligations.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
\seealso{
total.d2e lt.d2e debt.ratio
}
