% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.annuity.R
\name{fv.annuity}
\alias{fv.annuity}
\title{Estimate Future Value of an Annuity}
\usage{
fv.annuity(r, n, pmt, type = 0)
}
\arguments{
\item{r}{Discount rate, or the interest rate at which the amount will be compounded each period.}

\item{n}{Number of periods.}

\item{pmt}{Payment per period.}

\item{type}{Payments occur at the end of each period (type=0); payments occur at the beginning of each period (type=1).}
}
\value{
Future value of the annuity.
}
\description{
Estimates the future value of an annuity based on the discount rate, number of periods, payment per period, and payment type.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
fv.annuity(r=0.03, n=12, pmt=-1000)

}
\seealso{
fv
}
