% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.R
\name{fv}
\alias{fv}
\title{Estimate Future Value (FV)}
\usage{
fv(r, n, pv = 0, pmt = 0, type = 0)
}
\arguments{
\item{r}{Discount rate, or the interest rate at which the amount will be compounded each period.}

\item{n}{Number of periods.}

\item{pv}{Present value.}

\item{pmt}{Payment per period.}

\item{type}{Payments occur at the end of each period (type=0); payments occur at the beginning of each period (type=1).}
}
\value{
Future value (FV).
}
\description{
Estimates the future value (FV) based on the discount rate, number of periods, present value, payment per period, and payment type.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
\seealso{
fv.simple fv.annuity pv pmt n.period discount.rate
}
