% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.simple.R
\name{fv.simple}
\alias{fv.simple}
\title{Estimate Future Value of a Single Sum}
\usage{
fv.simple(r, n, pv)
}
\arguments{
\item{r}{Discount rate, or the interest rate at which the amount will be compounded each period.}

\item{n}{Number of periods.}

\item{pv}{Present value.}
}
\value{
Future value (FV) of the single sum.
}
\description{
Estimates the future value (FV) of a single sum based on the discount rate, number of periods, and present value.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
fv.simple(0.08, 10, -300)

}
\seealso{
fv
}
