% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.uneven.R
\name{fv.uneven}
\alias{fv.uneven}
\title{Computing the Future Value of an Uneven Cash Flow Series}
\usage{
fv.uneven(r, cf)
}
\arguments{
\item{r}{Stated annual rate.}

\item{cf}{Uneven cash flow series.}
}
\value{
Future value of the uneven cash flow series.
}
\description{
Computes the future value of an uneven cash flow series based on the stated annual rate and the cash flow series.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
\seealso{
fv.simple
}
