% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.R
\name{fv_single}
\alias{fv_single}
\title{Future Value of a Single Amount (FV Single) Function}
\usage{
fv_single(amount, rate, periods)
}
\arguments{
\item{amount}{The initial amount.}

\item{rate}{The discount rate.}

\item{periods}{The number of periods.}
}
\value{
The Future Value of the single amount.
}
\description{
Calculates the Future Value (FV) of a single amount given a discount rate and number of periods.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
fv_single(100, 0.1, 3)
}
