% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric.mean.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Geometric Mean Return}
\usage{
geometric.mean(r)
}
\arguments{
\item{r}{Returns over multiple periods.}
}
\value{
Geometric mean return.
}
\description{
Computes the geometric mean return based on returns over multiple periods.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
