% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ohlc.yahoo.R
\name{get.ohlc.yahoo}
\alias{get.ohlc.yahoo}
\title{Download Stock Prices from Yahoo Finance}
\usage{
get.ohlc.yahoo(symbol, start = "firstDay", end = "today", freq = "d")
}
\arguments{
\item{symbol}{Symbol of stock, e.g., AAPL, GOOG, SPX.}

\item{start}{Start date, e.g., 2013-01-01.}

\item{end}{End date, e.g., 2013-08-06.}

\item{freq}{Time interval, e.g., d: daily, w: weekly, m: monthly.}
}
\value{
Data frame with columns: Date, Open, High, Low, Close, Volume, Adjusted.
}
\description{
Downloads stock prices from Yahoo Finance (open, high, low, close, volume, adjusted).
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
# get.ohlc.yahoo(symbol="AAPL")
# get.ohlc.yahoo(symbol="AAPL", start="2013-08-01", freq="d")

}
\seealso{
get.ohlcs.yahoo get.ohlcs.google
}
