% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{ggm}
\alias{ggm}
\title{Gordon Growth Model (GGM) Function}
\usage{
ggm(dividend, growth_rate, discount_rate)
}
\arguments{
\item{dividend}{The current dividend per share.}

\item{growth_rate}{The expected constant growth rate of dividends.}

\item{discount_rate}{The discount rate (required rate of return).}
}
\value{
The estimated fair value of the stock.
}
\description{
Calculates the fair value of a stock using the Gordon Growth Model.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
}
