% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm.R
\name{gpm}
\alias{gpm}
\title{Gross Profit Margin}
\usage{
gpm(gp, rv)
}
\arguments{
\item{gp}{Gross profit, equal to revenue minus cost of goods sold (cogs).}

\item{rv}{Revenue (sales).}
}
\value{
Gross profit margin.
}
\description{
Computes the gross profit margin, evaluating a company’s financial performance.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:

}
