% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.mean.R
\name{harmonic.mean}
\alias{harmonic.mean}
\title{Harmonic Mean}
\usage{
harmonic.mean(p)
}
\arguments{
\item{p}{Price over multiple periods.}
}
\value{
Harmonic mean.
}
\description{
Computes the harmonic mean, average price over multiple periods.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica

Harmonic Mean, Average Price
}
\examples{
# Example usage:

}
