% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr.R
\name{hpr}
\alias{hpr}
\title{Holding Period Return}
\usage{
hpr(ev, bv, cfr = 0)
}
\arguments{
\item{ev}{Ending value.}

\item{bv}{Beginning value.}

\item{cfr}{Cash flow received.}
}
\value{
Holding period return.
}
\description{
Computes the holding period return (HPR) based on beginning and ending values and cash flows received.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica

#' Computing HPR, the Holding Period Return
}
\examples{
# Example usage:

}
\seealso{
twrr hpr2ear hpr2mmy
}
