% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr2bey.R
\name{hpr2bey}
\alias{hpr2bey}
\title{Bond-Equivalent Yield (BEY)}
\usage{
hpr2bey(hpr, t)
}
\arguments{
\item{hpr}{Holding period return.}

\item{t}{Number of months remaining until maturity.}
}
\value{
Bond-equivalent yield (BEY).
}
\description{
Computes the bond-equivalent yield (BEY), which is 2 times the semiannual discount rate.
}
\details{
\subsection{Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu}{

Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica

Bond-Equivalent Yield (BEY), 2 x the Semiannual Discount Rate
}
}
\examples{
# Example usage:

}
\seealso{
hpr
}
