% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpr2ear.R
\name{hpr2ear}
\alias{hpr2ear}
\title{Effective Annual Rate (EAR)}
\usage{
hpr2ear(hpr, t)
}
\arguments{
\item{hpr}{Holding period return.}

\item{t}{Number of days remaining until maturity.}
}
\value{
Effective annual rate (EAR).
}
\description{
Converts the holding period return (HPR) to the effective annual rate (EAR).
}
\details{
\subsection{Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu}{

Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica

#' Convert Holding Period Return to Effective Annual Rate
}
}
\examples{
# Example usage:

}
\seealso{
ear

hpr

ear2hpr
}
