% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate.R
\name{interest_rate}
\alias{interest_rate}
\title{Interest Rate Function}
\usage{
interest_rate(pv, fv, period)
}
\arguments{
\item{pv}{The present value.}

\item{fv}{The future value.}

\item{period}{The period over which the interest rate is calculated.}
}
\value{
The interest rate.
}
\description{
Calculates the interest rate given present value (PV) and future value (FV).
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
}
