% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr2.R
\name{irr2}
\alias{irr2}
\title{Internal Rate of Return (IRR)}
\usage{
irr2(cf, cutoff = 0.1, from = -1, to = 10, step = 0.000001)
}
\arguments{
\item{cf}{Cash flows. The first cash flow is the initial outlay (negative value).}

\item{cutoff}{Threshold to consider NPV as zero.}

\item{from}{Smallest IRR to try.}

\item{to}{Largest IRR to try.}

\item{step}{Increment of the IRR.}
}
\value{
Internal rate of return (IRR).
}
\description{
This function calculates the internal rate of return (IRR) for cash flows, including negative values.
}
\details{
\subsection{Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu}{

Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica

Computing IRR, the Internal Rate of Return
}
}
\examples{
# Example usage:
irr2(cf = c(-5, 1.6, 2.4, 2.8))

}
\seealso{
irr
}
