% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirr.R
\name{mirr}
\alias{mirr}
\title{Modified Internal Rate of Return (MIRR) Function}
\usage{
mirr(cash_flows, finance_rate, reinvest_rate)
}
\arguments{
\item{cash_flows}{A numeric vector of cash flows.}

\item{finance_rate}{The finance rate for discounting cash outflows.}

\item{reinvest_rate}{The reinvestment rate for compounding cash inflows.}
}
\value{
The Modified Internal Rate of Return.
}
\description{
Calculates the Modified Internal Rate of Return (MIRR) given cash flows, finance rate, and reinvestment rate.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
}
