% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetuity_pv.R
\name{perpetuity_pv}
\alias{perpetuity_pv}
\title{Present Value (PV) of a Perpetuity Function}
\usage{
perpetuity_pv(rate, pmt)
}
\arguments{
\item{rate}{The discount rate.}

\item{pmt}{The payment amount per period.}
}
\value{
The Present Value.
}
\description{
Calculates the Present Value (PV) of a perpetuity.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
perpetuity_pv(0.05, 100)
}
