% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_efficient_frontier.R
\name{plot_efficient_frontier}
\alias{plot_efficient_frontier}
\title{Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica}
\usage{
plot_efficient_frontier(frontier_data)
}
\arguments{
\item{frontier_data}{A data frame with columns Return, Risk, and Sharpe_Ratio.}
}
\value{
A ggplot2 object representing the Efficient Frontier.
}
\description{
Plot Efficient Frontier
}
\details{
This function plots the Efficient Frontier given a data frame containing Return, Risk, and Sharpe Ratio.
}
