% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_portfolio_optimization.R
\name{plot_portfolio_optimization}
\alias{plot_portfolio_optimization}
\title{Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
Plot Portfolio Optimization}
\usage{
plot_portfolio_optimization(returns, covariance_matrix)
}
\arguments{
\item{returns}{A matrix or data frame of asset returns.}

\item{covariance_matrix}{A covariance matrix of asset returns.}
}
\value{
A Plotly object representing the optimal portfolio weights.
}
\description{
This function plots the optimal portfolio weights given asset returns and covariance matrix.
}
\details{
First you need to calculate portfolio_optimization
}
\examples{
# Example usage:
}
