% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt.R
\name{pmt}
\alias{pmt}
\title{Period Payment (PMT)}
\usage{
pmt(r, n, pv, fv, type = 0)
}
\arguments{
\item{r}{Discount rate, or the interest rate at which the amount will be compounded each period.}

\item{n}{Number of periods.}

\item{pv}{Present value.}

\item{fv}{Future value.}

\item{type}{Payments occur at the end of each period (type = 0); payments occur at the beginning of each period (type = 1).}
}
\value{
Payment per period.
}
\description{
Estimates the payment per period for a loan or investment based on given parameters.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
pmt(r = 0.08, n = 10, pv = -1000, fv = 10)
pmt(r = 0.08, n = 10, pv = -1000, fv = 0)
pmt(r = 0.08, n = 10, pv = -1000, fv = 10, type = 1)

}
