% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_composition.R
\name{portfolio_composition}
\alias{portfolio_composition}
\title{Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica}
\usage{
portfolio_composition(weights, asset_names)
}
\arguments{
\item{weights}{A numeric vector of asset weights.}

\item{asset_names}{A character vector of asset names.}
}
\value{
A ggplot2 object representing the portfolio composition.
}
\description{
Plot Portfolio Composition
}
\details{
This function plots the composition of a portfolio given asset weights and names.
}
\examples{
weights <- c(0.3, 0.5, 0.2)
asset_names <- c("Stock A", "Stock B", "Stock C")
portfolio_composition(weights, asset_names)
}
