% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_optimization.R
\name{portfolio_optimization}
\alias{portfolio_optimization}
\title{Portfolio Optimization Function}
\usage{
portfolio_optimization(returns, covariance_matrix)
}
\arguments{
\item{returns}{A matrix or data frame of asset returns.}

\item{covariance_matrix}{A covariance matrix of asset returns.}
}
\value{
A numeric vector of optimal portfolio weights.
}
\description{
Optimizes portfolio weights to maximize return or minimize risk.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
}
