% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.annuity.R
\name{pv.annuity}
\alias{pv.annuity}
\title{Present Value of an Annuity (PV.Annuity)}
\usage{
pv.annuity(r, n, pmt, type = 0)
}
\arguments{
\item{r}{Discount rate, or the interest rate at which the amount will be compounded each period.}

\item{n}{Number of periods.}

\item{pmt}{Payment per period.}

\item{type}{Payments occur at the end of each period (type = 0); payments occur at the beginning of each period (type = 1).}
}
\value{
Present value (PV) of an annuity.
}
\description{
Estimates the present value (PV) of an annuity based on given parameters.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
pv.annuity(r = 0.03, n = 12, pmt = 1000)
pv.annuity(r = 0.0425, n = 3, pmt = 30000)

}
