% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Present Value (PV) Function}
\usage{
pv(rate, cash_flow, period)
}
\arguments{
\item{rate}{The discount rate.}

\item{cash_flow}{The cash flow amount.}

\item{period}{The period in which the cash flow occurs.}
}
\value{
The Present Value.
}
\description{
Calculates the Present Value (PV) of a single cash flow.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
}
