% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.uneven.R
\name{pv.uneven}
\alias{pv.uneven}
\title{Present Value of Uneven Cash Flows}
\usage{
pv.uneven(r, cf)
}
\arguments{
\item{r}{The discount rate.}

\item{cf}{A vector of cash flows.}
}
\value{
The present value of the cash flows.
}
\description{
Calculates the present value of a series of uneven cash flows.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
