% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.ratio.R
\name{quick.ratio}
\alias{quick.ratio}
\title{Quick Ratio}
\usage{
quick.ratio(cash, ms, rc, cl)
}
\arguments{
\item{cash}{The amount of cash.}

\item{ms}{Marketable securities.}

\item{rc}{Receivables.}

\item{cl}{Current liabilities.}
}
\value{
The quick ratio.
}
\description{
Liquidity ratios measure the firm’s ability to satisfy its short-term obligations as they come due.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
