% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.continuous.R
\name{r.continuous}
\alias{r.continuous}
\title{Convert Nominal Rate to Continuous Compounded Rate}
\usage{
r.continuous(r, m)
}
\arguments{
\item{r}{Nominal rate.}

\item{m}{Number of times compounded each year.}
}
\value{
The continuously compounded rate.
}
\description{
Converts a given nominal rate to a continuously compounded rate.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
