% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.norminal.R
\name{r.norminal}
\alias{r.norminal}
\title{Convert Continuous Compounded Rate to Nominal Rate}
\usage{
r.norminal(rc, m)
}
\arguments{
\item{rc}{Continuous compounded rate.}

\item{m}{Number of desired times compounded each year.}
}
\value{
The nominal rate.
}
\description{
Converts a given continuous compounded rate to a nominal rate.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
