% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.perpetuity.R
\name{r.perpetuity}
\alias{r.perpetuity}
\title{Rate of Return for a Perpetuity}
\usage{
r.perpetuity(pmt, pv)
}
\arguments{
\item{pmt}{Payment per period.}

\item{pv}{Present value.}
}
\value{
The rate of return.
}
\description{
Calculates the rate of return for a perpetuity.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
