% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf.R
\name{rcf}
\alias{rcf}
\title{Residual Cash Flow (RCF) Model Function}
\usage{
rcf(residual_cash_flows, discount_rate)
}
\arguments{
\item{residual_cash_flows}{A numeric vector of expected future residual cash flows.
Vector numérico de los flujos de efectivo residual esperados futuros.}

\item{discount_rate}{The discount rate (required rate of return).
Tasa de descuento (tasa requerida de retorno).}
}
\value{
The estimated fair value of the stock.
El valor estimado justo de la acción.
}
\description{
Calculates the present value of future residual cash flows using the Residual Cash Flow Model.

Esta función calcula el valor presente de los flujos de efectivo residual futuros utilizando el modelo de Residual Cash Flow.
}
\details{
Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
# Ejemplo de uso:

}
