% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.error.R
\name{sampling.error}
\alias{sampling.error}
\title{Computing Sampling Error}
\usage{
sampling.error(sm, mu)
}
\arguments{
\item{sm}{Sample mean.}

\item{mu}{Population mean.}
}
\value{
The sampling error.
}
\description{
Computes the sampling error.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
