% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFRatio.R
\name{SFRatio}
\alias{SFRatio}
\title{Computing Roy’s Safety-First Ratio}
\usage{
SFRatio(rp, rl, sd)
}
\arguments{
\item{rp}{Portfolio return.}

\item{rl}{Threshold level return.}

\item{sd}{Standard deviation of portfolio returns.}
}
\value{
Roy’s safety-first ratio.
}
\description{
Computes Roy’s safety-first ratio.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
