% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sharpe.ratio.R
\name{Sharpe.ratio}
\alias{Sharpe.ratio}
\title{Computing Sharpe Ratio}
\usage{
Sharpe.ratio(rp, rf, sd)
}
\arguments{
\item{rp}{Portfolio return.}

\item{rf}{Risk-free return.}

\item{sd}{Standard deviation of portfolio returns.}
}
\value{
The Sharpe Ratio.
}
\description{
Computes the Sharpe Ratio.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
