% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slde.R
\name{slde}
\alias{slde}
\title{Depreciation Expense Recognition – Straight-line Depreciation (SL)}
\usage{
slde(cost, rv, t)
}
\arguments{
\item{cost}{Cost of long-lived assets.}

\item{rv}{Residual value of the long-lived assets at the end of its useful life.}

\item{t}{Length of the useful life.}
}
\value{
The annual depreciation expense.
}
\description{
Allocates an equal amount of depreciation each year over the asset’s useful life.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
