% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.d2e.R
\name{total.d2e}
\alias{total.d2e}
\title{Total Debt-to-Equity Ratio}
\usage{
total.d2e(td, te)
}
\arguments{
\item{td}{Total debt.}

\item{te}{Total equity.}
}
\value{
The total debt-to-equity ratio.
}
\description{
Solvency ratios measure the firm’s ability to satisfy its long-term obligations.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
