% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twrr.R
\name{twrr}
\alias{twrr}
\title{Computing TWRR}
\usage{
twrr(ev, bv, cfr)
}
\arguments{
\item{ev}{Ordered ending value list.}

\item{bv}{Ordered beginning value list.}

\item{cfr}{Ordered cash flow received list.}
}
\value{
The time-weighted rate of return.
}
\description{
Computes the time-weighted rate of return (TWRR).
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
