% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeChart.R
\name{volumeChart}
\alias{volumeChart}
\title{Volume Chart}
\usage{
volumeChart(ohlc, main = "", ...)
}
\arguments{
\item{ohlc}{Output from get.ohlc.yahoo or get.ohlc.google.}

\item{main}{An overall title for the plot.}
}
\value{
A ggplot object showing the volume chart.
}
\description{
Shows each period’s volume as a vertical line.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
