% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/was.R
\name{was}
\alias{was}
\title{Calculate Weighted Average Shares}
\usage{
was(ns, nm)
}
\arguments{
\item{ns}{n x 1 vector of number of shares.}

\item{nm}{n x 1 vector of number of months related to ns.}
}
\value{
The weighted average number of common shares.
}
\description{
Calculate weighted average shares – weighted average number of common shares.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
s <- c(10000, 2000)
m <- c(12, 6)
was(ns = s, nm = m)
}
