% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpr.R
\name{wpr}
\alias{wpr}
\title{Weighted Mean as a Portfolio Return}
\usage{
wpr(r, w)
}
\arguments{
\item{r}{Returns of the individual assets in the portfolio.}

\item{w}{Corresponding weights associated with each of the individual assets.}
}
\value{
The weighted mean portfolio return.
}
\description{
Calculate the weighted mean as a portfolio return.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
}
