% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_yield_plotly.R
\name{bond_yield_plotly}
\alias{bond_yield_plotly}
\title{Bond Yield Function Plot using Plotly}
\usage{
bond_yield_plotly(face_value, price, coupon_rate, nper)
}
\arguments{
\item{face_value}{The face value (par value) of the bond.}

\item{price}{The current market price of the bond.}

\item{coupon_rate}{The annual coupon rate of the bond.}

\item{nper}{The number of periods (years) until bond maturity.}
}
\value{
A plotly object representing the bond yield function.
}
\description{
This function plots the bond yield function given the bond parameters.
}
\details{
Professor of PhD and Master's in Systems, Economics, and Statistics for Scientific Research
}
\examples{
# Example 1: Plot the bond yield function for a bond with a face value of $1000, current price $950,
#            5\% coupon rate, and 10 years until maturity.

# Example 2: Plot the bond yield function for a bond with a face value of $500, current price $480,
#            4.5\% coupon rate, and 5 years until maturity.
}
