% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca.R
\name{cca}
\alias{cca}
\title{Comparable Companies Analysis (CCA) Function}
\usage{
cca(comparable_pe_ratio, earnings_per_share)
}
\arguments{
\item{comparable_pe_ratio}{The average Price-to-Earnings (P/E) Ratio of comparable companies.}

\item{earnings_per_share}{The current earnings per share.}
}
\value{
The estimated fair value of the stock.
}
\description{
Estimates the fair value of a stock using Comparable Companies Analysis.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica

Comparable Companies Analysis (CCA) Function
}
\examples{
# Example usage:
}
