% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddgm.R
\name{ddgm}
\alias{ddgm}
\title{Discounted Dividend Growth Model (DDGM) Function}
\usage{
ddgm(dividend, growth_rate, discount_rate)
}
\arguments{
\item{dividend}{The current dividend per share.}

\item{growth_rate}{The expected constant growth rate of dividends.}

\item{discount_rate}{The discount rate (required rate of return).}
}
\value{
The estimated fair value of the stock.
}
\description{
Calculates the fair value of a stock using the Discounted Dividend Growth Model.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
dividend <- 2.5
growth_rate <- 0.04
discount_rate <- 0.1
ddgm(dividend, growth_rate, discount_rate)
}
