% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_price.R
\name{dirty_price}
\alias{dirty_price}
\title{Calculate Dirty Price}
\usage{
dirty_price(clean_price, accrued_interest)
}
\arguments{
\item{clean_price}{The clean price of the bond.}

\item{accrued_interest}{The accrued interest of the bond.}
}
\value{
The dirty price of the bond.
}
\description{
Calculates the dirty price of a bond given the clean price and the accrued interest.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
dirty_price(100, 2)
}
