% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva.R
\name{eva}
\alias{eva}
\title{Economic Value Added (EVA) Model Function}
\usage{
eva(net_operating_profit_after_tax, weighted_average_cost_of_capital)
}
\arguments{
\item{net_operating_profit_after_tax}{Ganancia Operativa Neta después de Impuestos (NOPAT).}

\item{weighted_average_cost_of_capital}{Costo Promedio Ponderado de Capital (WACC).}
}
\value{
The estimated fair value of the stock.

El valor estimado justo de la acción.
}
\description{
Calculates the Economic Value Added (EVA) for a stock.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
net_operating_profit_after_tax <- 1000000
weighted_average_cost_of_capital <- 0.08
eva(net_operating_profit_after_tax, weighted_average_cost_of_capital)

# Ejemplo de uso:

}
