% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_rate.R
\name{interest_rate}
\alias{interest_rate}
\title{Interest Rate Function}
\usage{
interest_rate(pv, fv, period)
}
\arguments{
\item{pv}{El valor presente.}

\item{fv}{El valor futuro.}

\item{period}{El periodo sobre el cual se calcula la tasa de interés.}
}
\value{
The interest rate.

La tasa de interés.
}
\description{
Calculates the interest rate given present value (PV) and future value (FV).

This function calculates the interest rate given the present value (PV) and future value (FV) over a specified period.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica
}
\examples{
# Example usage:
interest_rate(100, 200, 5)
# Ejemplo de uso:
}
