% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_efficient_frontier_B.R
\name{plot_efficient_frontier_B}
\alias{plot_efficient_frontier_B}
\title{Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Científica}
\usage{
plot_efficient_frontier_B(frontier_data, max_sharpe_data = NULL)
}
\arguments{
\item{frontier_data}{A data frame with columns Return, Risk, and Sharpe_Ratio.
Un marco de datos con las columnas Return, Risk y Sharpe_Ratio.}
}
\value{
A ggplot2 object representing the Efficient Frontier.
Un objeto ggplot2 que representa la Frontera Eficiente.
}
\description{
Plot Efficient Frontier using ggplot2
}
\details{
This function plots the Efficient Frontier given a data frame containing Return, Risk, and Sharpe Ratio using ggplot2.
}
\examples{
# Example usage with efficient_frontier function
returns <- matrix(rnorm(100 * 4, mean = 0.01, sd = 0.02), nrow = 100, ncol = 4)
colnames(returns) <- c("AMZN", "MSFT", "TSLA", "AAPL")
cov_matrix <- cov(returns)
frontier <- efficient_frontier(returns, cov_matrix)
plot_efficient_frontier_B(frontier)

}
