% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_ratio.R
\name{pre_ratio}
\alias{pre_ratio}
\title{Price-to-Earnings (P/E) Ratio Model Function}
\usage{
pre_ratio(earnings_per_share, pe_ratio)
}
\arguments{
\item{earnings_per_share}{The current earnings per share.}

\item{pe_ratio}{The desired Price-to-Earnings (P/E) Ratio.}
}
\value{
The estimated fair value of the stock.
}
\description{
Estimates the fair value of a stock using the Price-to-Earnings (P/E) Ratio.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
}
