% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf.R
\name{rcf}
\alias{rcf}
\title{Residual Cash Flow (RCF) Model Function}
\usage{
rcf(residual_cash_flows, discount_rate)
}
\arguments{
\item{residual_cash_flows}{A numeric vector of expected future residual cash flows.}

\item{discount_rate}{The discount rate (required rate of return).}
}
\value{
The estimated fair value of the stock.
}
\description{
Calculates the present value of future residual cash flows using the Residual Cash Flow Model.
}
\details{
#'#' Dr. Roberto Chang López  rchang@unah.edu.hn / rchang@unitec.edu
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
#'
}
\examples{
# Example usage:
residual_cash_flows <- c(5, 5.5, 6, 6.5, 7)
discount_rate <- 0.1
rcf(residual_cash_flows, discount_rate)
}
