% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasa_indiferencia_lempiras.R
\name{tasa_indiferencia_lempiras}
\alias{tasa_indiferencia_lempiras}
\title{Indifference Rate Function (Lempiras)}
\usage{
tasa_indiferencia_lempiras(tasa_en_dolares, devaluacion_anual)
}
\arguments{
\item{tasa_en_dolares}{The dollar interest rate.}

\item{devaluacion_anual}{The annual devaluation rate of Lempira.}

\item{tasa_en_dolares_anual}{The effective annual dollar interest rate.}
}
\value{
The indifference rate in Lempiras.
}
\description{
Calculates the indifference rate in Lempiras given a dollar interest rate, annual devaluation rate of Lempira, and optionally a constant.
}
\details{
Profesor de Doctorado y Maestría en Sistemas, Economía y Estadística para la Investigación Ciéntífica
}
\examples{
# Example usage:
}
